split adapter tp primary and secondary package
