--- adapter/src/main/scala/librame/adapter/primary/JsValueError.scala
@@ -1,16 +1,0 @@
-package librame.adapter.primary
-
-import play.api.libs.json._
-
-/**
- * @param errorCode
- * @param target
- */
-case class JsValueError(
-  errorCode: String,
-  target:    String
-)
-
-object JsValueError {
-  implicit val errorWrites: Writes[JsValueError] = Json.writes[JsValueError]
-}
