--- adapter/src/main/scala/librame/adapter/secondary/Exception.scala
@@ -1,7 +1,0 @@
-package librame.adapter.secondary
-
-import com.mysql.jdbc.exceptions.jdbc4.MySQLIntegrityConstraintViolationException
-
-object exception {
-  type ConstraintException = MySQLIntegrityConstraintViolationException
-}
