--- adapter/src/main/scala/librame/adapter/secondary/mail/MailerRepository.scala
@@ -1,13 +1,0 @@
-package librame.adapter.secondary.mail
-
-import scala.concurrent.Future
-
-import librame.domain.model.EmailAddress
-
-trait MailerRepository {
-  /**
-   * @param to
-   * @return
-   */
-  def send(to: EmailAddress): Future[Option[String]]
-}
