--- adapter/src/main/scala/librame/adapter/secondary/mail/MockMailerRepository.scala
@@ -1,24 +1,0 @@
-package librame.adapter.secondary.mail
-
-import java.util.UUID
-
-import scala.concurrent.Future
-
-import librame.domain.model.EmailAddress
-
-class MockMailerRepository extends MailerRepository {
-  /**
-   * @param to
-   * @return
-   */
-  def send(to: EmailAddress): Future[Option[String]] = {
-    val token = UUID.randomUUID.toString
-
-    Future.successful(
-      token.isEmpty match {
-        case true  => None
-        case false => Some(token)
-      }
-    )
-  }
-}
