--- adapter/src/main/scala/librame/adapter/secondary/rdb/DoobieRepository.scala
@@ -1,7 +1,0 @@
-package librame.adapter.secondary.rdb
-
-import play.api.db.DBApi
-
-class DoobieRepository(db: DBApi)
-  extends DoobieDatabaseConfig(db)
-  with DoobieCustomColumnType
