--- adapter/src/main/scala/librame/adapter/secondary/rdb/SlickDatabaseConfig.scala
@@ -1,13 +1,0 @@
-package librame.adapter.secondary.rdb
-
-import slick.jdbc.JdbcBackend.Database
-
-
-trait SlickDatabaseConfig {
-  val db: Database = Database.forConfig("rdb")
-
-  val profile = slick.jdbc.MySQLProfile
-}
-
-object SlickDatabaseConfig extends SlickDatabaseConfig
-
