--- adapter/src/main/scala/librame/adapter/secondary/rdb/SlickRepository.scala
@@ -1,7 +1,0 @@
-package librame.adapter.secondary.rdb
-
-import scala.concurrent.ExecutionContext
-
-trait SlickRepository extends SlickDatabaseConfig with SlickCustomColumnType {
-  implicit val ec: ExecutionContext = ExecutionContext.global
-}
