--- adapter/src/main/scala/librame/adapter/secondary/rdb/SlickTable.scala
@@ -1,14 +1,0 @@
-package librame.adapter.secondary.rdb
-
-import SlickDatabaseConfig.profile
-import slick.lifted.Tag
-
-
-/**
- * @param tag
- * @param tableName
- * @tparam M
- */
-abstract class SlickTable[M](tag: Tag, tableName: String)
-  extends profile.Table[M](tag, tableName)
-  with SlickCustomColumnType
