+++ secondary-adapter/src/main/scala/librame/secondary/adapter/Exception.scala
@@ -1,0 +1,7 @@
+package librame.secondary.adapter
+
+import com.mysql.jdbc.exceptions.jdbc4.MySQLIntegrityConstraintViolationException
+
+object exception {
+  type ConstraintException = MySQLIntegrityConstraintViolationException
+}
