+++ secondary-adapter/src/main/scala/librame/secondary/adapter/mail/MailerRepository.scala
@@ -1,0 +1,13 @@
+package librame.secondary.adapter.mail
+
+import scala.concurrent.Future
+
+import librame.domain.model.EmailAddress
+
+trait MailerRepository {
+  /**
+   * @param to
+   * @return
+   */
+  def send(to: EmailAddress): Future[Option[String]]
+}
