+++ secondary-adapter/src/main/scala/librame/secondary/adapter/rdb/DoobieRepository.scala
@@ -1,0 +1,7 @@
+package librame.secondary.adapter.rdb
+
+import play.api.db.DBApi
+
+class DoobieRepository(db: DBApi)
+  extends DoobieDatabaseConfig(db)
+  with DoobieCustomColumnType
