+++ secondary-adapter/src/main/scala/librame/secondary/adapter/rdb/SlickDatabaseConfig.scala
@@ -1,0 +1,13 @@
+package librame.secondary.adapter.rdb
+
+import slick.jdbc.JdbcBackend.Database
+
+
+trait SlickDatabaseConfig {
+  val db: Database = Database.forConfig("rdb")
+
+  val profile = slick.jdbc.MySQLProfile
+}
+
+object SlickDatabaseConfig extends SlickDatabaseConfig
+
