+++ secondary-adapter/src/main/scala/librame/secondary/adapter/rdb/SlickRepository.scala
@@ -1,0 +1,7 @@
+package librame.secondary.adapter.rdb
+
+import scala.concurrent.ExecutionContext
+
+trait SlickRepository extends SlickDatabaseConfig with SlickCustomColumnType {
+  implicit val ec: ExecutionContext = ExecutionContext.global
+}
